Rename conf files to local-storage
