+++ packages/app/src/cli/services/info.ts
@@ -1,5 +1,5 @@
 import {outputEnv} from './app/env/show.js'
-import {CachedAppInfo, getAppInfo} from './conf.js'
+import {CachedAppInfo, getAppInfo} from './local-storage.js'
 import {AppInterface} from '../models/app/app.js'
 import {FunctionExtension, ThemeExtension, UIExtension} from '../models/app/extensions.js'
 import {configurationFileNames} from '../constants.js'
