+++ packages/app/src/cli/services/local-storage.test.ts
@@ -1,4 +1,4 @@
-import {AppLocalStorageSchema, clearAppInfo, getAppInfo, setAppInfo} from './storage.js'
+import {AppLocalStorageSchema, clearAppInfo, getAppInfo, setAppInfo} from './local-storage.js'
 import {describe, expect, it} from 'vitest'
 import {LocalStorage} from '@shopify/cli-kit/node/local-storage'
 import {inTemporaryDirectory} from '@shopify/cli-kit/node/fs'
