+++ packages/theme/src/cli/services/list.test.ts
@@ -1,6 +1,6 @@
 import {list} from './list.js'
 import {columns} from './list.columns.js'
-import {getDevelopmentTheme} from './conf.js'
+import {getDevelopmentTheme} from './local-storage.js'
 import {fetchStoreThemes} from '../utilities/theme-selector/fetch.js'
 import {Theme} from '@shopify/cli-kit/node/themes/models/theme'
 import {renderTable} from '@shopify/cli-kit/node/ui'
@@ -8,7 +8,7 @@ import {describe, expect, it, vi} from 'vitest'
 
 vi.mock('../utilities/theme-selector/fetch.js')
 vi.mock('@shopify/cli-kit/node/ui')
-vi.mock('./conf.js')
+vi.mock('./local-storage.js')
 
 const session = {
   token: 'token',
